Fixed item context menus appearing after mouse has exited the item area.
This occurred because the scene does not receive mouse move events while a context menu is displayed.
If the user right-clicks on a new location while the menu is open, then the click event is delieverd as if the mouse had not moved.

Corrected by sending a just-in-time hover event immediately before mouse press, if the cursor has moved.